+++ test/src/Repository/RecipeRepositoryTest.php
@@ -502,8 +502,7 @@ class RecipeRepositoryTest extends TestCase
                    ->method('removeIds')
                    ->with($orphanedIds);
 
-        $result = $repository->removeOrphans();
-        $this->assertSame($repository, $result);
+        $repository->removeOrphans();
     }
 
     /**
@@ -662,7 +661,6 @@ class RecipeRepositoryTest extends TestCase
                    ->method('getEntityName')
                    ->willReturn($entityName);
 
-        $result = $this->invokeMethod($repository, 'removeIds', $recipeIds);
-        $this->assertSame($repository, $result);
+        $this->invokeMethod($repository, 'removeIds', $recipeIds);
     }
 }