+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Unreleased
 
+###
+
+- Added `RepositoryWithOrphansInterface` to signal that orphans must be cleaned. 
+
 ### Removed
 
 - `EntityType` interface. It has been replaced by the interface from `factorio-item-browser/common`.
+- Fluent interface from repository methods.
 
 ## 1.0.0 - 2018-08-04